Refactor logout view out of AccountView
